fix(agent): refine delivery artifact handling
